refactor: read search params server-side
